movies/fastapi/pyproject.toml
[tool.poetry]
name = "fast-api-1"
version = "0.1.0"
description = "First FastAPI sprint"
authors = ["Maksim Kosinov <mkosinov@mail.ru>", "Борис Гельруд <agu-66@mail.ru>"]
license = "open"
readme = "readme.md"

[tool.poetry.dependencies]
python = "^3.11"
redis = "^5.0.1"
elasticsearch = {version = "7.9.1", extras = ["async"]}
fastapi = "^0.104.1"
pydantic = "^2.5.2"
uvicorn = "^0.24.0.post1"
uvloop = "^0.19.0"
orjson = "^3.9.10"
pydantic-settings = "^2.1.0"
gunicorn = "^21.2.0"
aiologger = "^0.7.0"
aiofiles = "^0.7.0"
python-jose = "^3.3.0"
argon2-cffi = "^23.1.0"
pycryptodomex = "^3.20.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
